test: consolidate mapper constructor-arg tests into one package-agnostic check
The prior commit closes the last remaining pocket of unnamed constructor
args anywhere in db/rdbms's mapper directory. With write.domain,
search.entities, read.domain, and read.replication all now fully named
(and zero remaining implicit resultType= usage anywhere), the reason
the two existing tests were scoped by package prefix -- staged rollout
across packages -- no longer applies.

Replaces RdbmsDbModelConstructorArgsMustBeNamedTest,
SearchEntityConstructorArgsMustBeNamedTest, and the
MapperConfigurationTestSupport they shared with one
RdbmsMapperConstructorArgsMustBeNamedTest that checks every resultMap/
statement in the mapper directory with no package filtering, so any
future package is covered automatically with zero new test needed.

The implicit-resultType= check needed a new guard once the package
filter was removed: without it, every <insert>/<update>/<delete>
statement's implicit void result type, and every scalar resultType=
(e.g. java.lang.Long for a COUNT(*)), registered as a synthetic
'-Inline' resultMap and produced false-positive failures. Filters on
resultMap.getType().isRecord() instead -- only records have the
positional-automapping risk this check targets.

Noted directly in the class Javadoc: isRecord() is a proxy for the risk,
not the risk itself -- MyBatis's positional fallback isn't actually
record-specific. This doesn't leave a live gap for write.domain/
search.entities today, since RdbmsDbModelMustBeRecordArchTest/
SearchEntityArchTest independently enforce those two packages stay
records; read.domain/read.replication have no equivalent ArchUnit guard
(though everything mapped there today is already a record).

Both invariants sabotage-tested directly on the consolidated check: a
name= typo and a reverted resultMap=-to-resultType= both correctly fail,
restored afterward. Verified no logic was silently dropped in the merge
-- the deleted classes were thin, self-contained wrappers with no extra
assertions beyond what's now inlined.
